refactor(svelte): more clean up
