+++ apps/playground-sveltekit/src/app.d.ts
@@ -1,13 +1,16 @@
 /// <reference types="@sveltejs/kit" />
+import type { Session as NextAuthSession } from "next-auth"
+import type { JWT } from "next-auth/jwt"
 
 // See https://kit.svelte.dev/docs/typescript
 // for information about these interfaces
 declare namespace App {
-  interface Locals {}
-
+  interface Locals {
+    token: JWT
+  }
   interface Platform {}
 
-  interface Session {}
+  interface Session extends NextAuthSession {}
 
   interface Stuff {}
 }
