+++ apps/playground-sveltekit/src/hooks.ts
@@ -1,0 +1,4 @@
+import { hooks } from "$lib/hooks"
+import { authOptions } from "./routes/api/auth/[...nextauth]"
+
+export const { handle, getSession } = hooks(authOptions)
